implemented dialogue between server and client
I made the server wait for a message from the client, which is .
terminated, and which should either be "distru gob." or
"distru json." which will serve either a binary gob or json en-
coded representation of the current Index.

I also improved server and client logging, though the formatting
could use a little more work.