added webserver and placeholder html
A webserver now runs on port 9048, and serves a placeholder page. It can be
modified in the future to serve additional pages, as well as proper html. The
matter of how to load the html may still be problematic.

I also removed some broken code for handling the HTTP requests.