pages are fundamentally more tree-friendly
I removed the Internals field for pages, and replaced it with a more
complex constructor, which returns a map[string]bool (to mimick a Set)
containing **absolute** links to other internal pages. If the links are
not absolute, then they are resolved using join(), which is defined in
index.go.