added recursive indexing
This is an extremely important feature. I reworked the page and site
construction code so as to allow recursive indexing. When distru begins
to index a site, it will index the main page, and use the resulting
internal links to get a list of other pages to also index. This is done
until the entire public directory tree of the site is discovered and
scraped for all of the external links.

closes #11