started new goroutine for index building
Distru used to invoke NewIndex() on startup to initialize the variable
Idx. This meant that that function had to be run before any other
pieces of the software could initialize. I just moved this initializ-
ation to a new goroutine, which is invoked inside of Serve(). This
allows the server to start first, and then for the index to build while
it is running.

The server will only serve only an empty index until it finishes invoking the constructor.

ref #12