restructured index maintaining for multithreading
Previously, Indexes were very nondynamic structures, for which there
existed one constructor, which would index a list of sites in series
(a very slow process) using only one thread.

This commit changes that process entirely, so that indexes are instead
maintained by a number of constantly running processes, Indexers. These
read from a common channel, owned by Serve(), that controls what sites
should be indexed. Once put into the channel, they are addressed by the
first free Indexer, which immediately removes them from the channel and
begins to index the site. When it finishes, it looks for any new items
in the channel, and so on. If the channel is closed with close(c), the
Indexers tied to that channel are terminated.

The number of indexers to run is passed as an argument to MaintainIndex(),
so the number of sites that can be indexed concurrently is settable.
Furthermore, target urls may be added at any time, by any process which
knows about the queue channel. This may in the future be tied to a user
interface.

ref #12