created on-demand indexing framework
I added another (TEMPORARY) server command, "distru index" for adding
new sites to the queue to be indexed. All it does is add whatever string
that is entered to the queue. It is then properly picked up and indexed
by the Indexers.

I also changed the original queue, a local variable in Serve() to a
global one declared in distru.go, Queue. I'm looking into making the
access channel a part of the Index object itself, so as to simplify the
use of indexes.