made Index contain its Indexer control chan
I moved the previously-global Queue chan into a field of Index. This
simplifies control of indexes, because functions no longer have to pass
the control channel as well as the Index.

Index.Queue is tagged with json:"-", to omit it from JSON encoding this
is important, because encoding/json cannot seem to handle the encoding
of chan objects.