added robots.txt compliance
I imported a robotstxt package from GitHub. I added checks to make sure
that Distru is allowed access to certain files. Eventally, we can add
support for Sitemap parts.

I defined a BotName for Distru: "Distru"

I refactored the type declarations in index.go, so as to put them
towards the top of the file.

Sites must now *not* be prepended with "http://"