cleaned up indexing command handling
When a user connects to a Distru instance and types "distru index",
then gives a site name, the Distru instance will put the site name in
the Idx.Queue control channel. This commit causes the connection to the
user to be closed before entering the site into the control channel.