improved the redirect stability and compatibility
I changed the javascript redirect to be able to support running Distru
with a non-root path. Previously, it directed to "/search/<term>", but
now it adds "search/<term>" to the previous url. This means that if
an Apache server grants access to the Distru server via a proxy, it
will not break.

closes #16