updated references to site types to use pointers
This was caused by a desire for better error detection and handling by
Indexers, and generally by the program. When newSite() fails for
whatever reason, be it because of lack of site permission, malformed or
empty pages, it will return nil, and the Indexers will report that
error, and ignore it.

This protects against malformed index commands (such as dropping a full
URI, including http://, into the Queue.)

closes #19