added rudimentary searching
* Bumped Version *

I added Index.Search(), which returns a non-relevance-weighted list of
pages that are applicable to the given terms. It does not sort by
number of matched words, or other factors. It does a mindless content
search using existing word counts.

I also reworked the page type to maintain its own fully qualified link,
and updated getPage().

Furthermore, I added a "distru search" command, which takes a raw,
single string to search, and produces the number of results in the
server logs.

ref #20