added search response
When a server recieves a SEARCH request from a connection, it will do
the search, record the number of results and the search term in the
server log, then write all of the result URLs to the connection, then
close it.

This is a great improvement towards interactive searching, though it
should still be implemented in the web interface.

ref #20