Removed cap on number of posts.
Before, when we were still testing it, there was a cap of 10 posts when searching.
Now there is no limit and as of right now if there are one million results,
it will show one million results. In the future there will be either
different pages or another way to show more posts, but for now this is a
much better way than caping the posts and only showing a few.