configuration addition and revamp
I added a relatively large update which restructures the core functions
to rely on the new config type, which contains options (which are for
the moment blank.) This can be saved to the location specified in
ConfPath (/etc/distru.conf) by sending the RPC server "distru save". It
is loaded on startup.

If the configuration file is not present or not readable, a new config
object is created on startup, and tagged with the current version
number. A config file is JSON-encoded and stores the entire index.

closes #2