added config-based autoindexing
I added an AutoIndex field in the config type to be read on startup,
which contains a list of sites to automatically index on startup (and in
the future, repetetively at specific intervals.)

Additionally, I updated Serve() to automatically queue these while
starting up. It does not pause execution of other startup routines,
though.

closes #36