added 4 second timeout on Index.MergeRemote()
I changed the net.Dial() function to net.DialTimeout() and forced the
connection to close after four seconds. If it times out, then the Index
is left as it is. This means that any WebUI search will not be stopped
for more than four seconds, waiting for Resources to respond with their
Indexes.

closes #45