fixed apparent pool deadlock
The original "pool" size for a single site index routine was 16. This
meant that 16 target site pages could be held at a time. Then, if this
ever became full, more pagers would be started, theoretically. (This
duplication behavior has now been disabled.) This seemed to cause
situations which looked like (but were not actually) thread deadlocks. I
did this by changing the pool size to 64.

I also disallowed "?" in URLs, because URLs containing them are
generally not good to search. In some cases, they are, but currently
cause more strain than is acceptable. This may be an issue.