switched to bencode (compatibility breaking)
I switched all encoding of Indexes (and sub-objects) to Bencode. This
will break all interaction between this version and previous versions.

That said, it should work better than JSON encoding, and fixes other
issues regarding that. It is slimmer and easier than JSON, and
significantly better suited to the task.

closes #64
closes #68