Created the mathematical portion of the upcoming features.
The next features will be a dictionary (if only one word is searched, define it!),
and more complex mathematical equations. Right now it does basic math, but more advanced
math can easily be done, too. Also, anything else we want to parse for other things will
be able to be done as well.