Reduced the size of parse.js.
Knowingly, I had a lot of extra code that didn't need to be duplicated,
but it was. I was only doing that at first just to test each part, but
now I know it's safe to put it all together. So now there is less code,
but it does the same thing.