Added support for exponents.
Exponents can now be used in the following bases:
2, 8, 10, 16.
Also fixed some errors where when doing simple math with
base 10 and any other base, it would think it was in that
base and convert it down to base 10, giving it either a
larger or smaller value depending on the base. Probably
fixed some other things, too, but I'm not quite sure off
the top of my head.

Ref #74.