Add --genconf and --confloc flags
I am adding flags to Distru, including --genconf, which saves the
default configuration, and --confloc, which prints the default
configuration location, (which has always been "/etc/distru.conf".)

When Distru is run with a filename as the first argument, that is
substituted for the default path. So, if one runs "distru
./distru.conf", distru will load from ./distru.conf instead of
/etc/distru.conf. Additionally, if one runs "distru --genconf
./distru.conf", the file ./distru.conf will be saved as the default
configuration file.